scripts/requirements.txt
matplotlib
numpy
pandas
PyYAML